# openapi.yaml
openapi: 3.0.3
info:
  title: Platform Marketplace Responden API
  version: 0.1.0
  description: >
    API untuk platform marketplace responden. Lihat PRD.md dan skill.md
    untuk aturan bisnis lengkap di balik setiap endpoint.

servers:
  - url: https://api.staging.example.com/v1
    description: Staging
  - url: https://api.example.com/v1
    description: Production

tags:
  - name: Auth
  - name: Profile
  - name: Research
  - name: Participation
  - name: Wallet
  - name: Withdrawal
  - name: Admin
  - name: Support

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrasi akun baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, phone, password, age_declared_18plus]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                password: { type: string, format: password }
                age_declared_18plus: { type: boolean, description: "Wajib true, self-declaration umur 18+" }
      responses:
        '201': { description: Akun dibuat, OTP dikirim }
        '400': { description: Validasi gagal }

  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verifikasi OTP email/HP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target, code]
              properties:
                target: { type: string, enum: [email, phone] }
                code: { type: string }
      responses:
        '200': { description: Terverifikasi }

  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: Login sukses, kembalikan JWT }

  /profile:
    get:
      tags: [Profile]
      summary: Ambil profil user saat ini
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    put:
      tags: [Profile]
      summary: Update profil (termasuk consent agama & data-share)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileInput'
      responses:
        '200': { description: Profil terupdate }

  /profile/domicile-verify:
    post:
      tags: [Profile]
      summary: Verifikasi domisili via GPS
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lat: { type: number }
                lng: { type: number }
      responses:
        '200': { description: Domisili terverifikasi }

  /research:
    get:
      tags: [Research]
      summary: List research tersedia (auto-filter sesuai profil respondent)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ResearchSummary' }
    post:
      tags: [Research]
      summary: Buat research baru (Researcher)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResearchInput' }
      responses:
        '201': { description: Research dibuat (status draft) }
        '400': { description: "Validasi gagal, mis. target_respondent_count < 50" }

  /research/{id}:
    get:
      tags: [Research]
      summary: Detail research
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: OK }

  /research/{id}/publish:
    post:
      tags: [Research]
      summary: Publish research (auto, tanpa review admin — validasi token saja)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Research published }
        '402': { description: Token tidak cukup }

  /research/{id}/cancel:
    post:
      tags: [Research]
      summary: Cancel research (sisa slot direfund dalam bentuk token)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Research dibatalkan, refund diproses }

  /research/{id}/export:
    get:
      tags: [Research]
      summary: Export hasil research ke CSV (termasuk data respondent)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: File CSV
          content:
            text/csv: {}

  /research/{id}/screening:
    post:
      tags: [Participation]
      summary: Submit jawaban screening question
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: "Lolos/gagal screening (berdasar skor/threshold)" }

  /research/{id}/participate:
    post:
      tags: [Participation]
      summary: Mulai partisipasi (setelah lolos screening)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Participation dibuat, status in_progress }

  /research/{id}/submit:
    post:
      tags: [Participation]
      summary: Submit jawaban survey (trigger auto-screening)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: "Jawaban diterima, masuk proses quality control" }

  /wallet:
    get:
      tags: [Wallet]
      summary: Lihat saldo token (dihitung dari ledger)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance_token: { type: integer }
                  balance_idr_equivalent: { type: number }

  /wallet/topup:
    post:
      tags: [Wallet]
      summary: Top up token via payment gateway
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_idr, payment_method, idempotency_key]
              properties:
                amount_idr: { type: number }
                payment_method: { type: string, enum: [qris, ewallet, va] }
                idempotency_key: { type: string }
      responses:
        '200': { description: "Redirect/instruksi pembayaran" }

  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: Riwayat transaksi token (ledger)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /payment/webhook:
    post:
      tags: [Wallet]
      summary: Webhook callback dari payment gateway (idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Diproses }

  /withdrawals:
    get:
      tags: [Withdrawal]
      summary: Riwayat withdrawal user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      tags: [Withdrawal]
      summary: Request withdrawal (potongan 3%)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token_amount]
              properties:
                token_amount: { type: integer }
      responses:
        '201': { description: "Withdrawal diajukan, status requested" }

  /admin/quality-review:
    get:
      tags: [Admin]
      summary: "Antrian jawaban yang di-flag (Admin Quality), SLA 48 jam"
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /admin/quality-review/{participationId}/decision:
    post:
      tags: [Admin]
      summary: Keputusan admin (approve/reject) untuk participation yang di-flag
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: participationId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                note: { type: string }
      responses:
        '200': { description: "Reward diproses (approved) atau dibatalkan (rejected)" }

  /admin/withdrawals/{id}/approve:
    post:
      tags: [Admin]
      summary: Approve withdrawal (Admin Finance)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Withdrawal diproses ke payment gateway }

  /admin/content-reports:
    get:
      tags: [Admin]
      summary: "Antrian report konten (Admin Quality), SLA 48 jam"
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  /support/tickets:
    post:
      tags: [Support]
      summary: Buat tiket support/banding
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, subject, description]
              properties:
                category: { type: string, enum: [appeal_respondent, appeal_researcher, other] }
                subject: { type: string }
                description: { type: string }
      responses:
        '201': { description: Tiket dibuat }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserProfileInput:
      type: object
      properties:
        gender: { type: string }
        religion: { type: string }
        religion_consent: { type: boolean, description: "Wajib true sebelum religion disimpan" }
        domicile_province: { type: string }
        domicile_city: { type: string }
        data_share_consent: { type: boolean, description: "Consent data dibagikan ke researcher via export" }

    ResearchInput:
      type: object
      required: [title, description, external_survey_url, target_respondent_count, estimated_duration_minutes, deadline]
      properties:
        title: { type: string }
        description: { type: string }
        external_survey_url: { type: string, format: uri }
        target_respondent_count: { type: integer, minimum: 50 }
        criteria:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              operator: { type: string }
              value: { type: string }
        screening_questions:
          type: array
          items:
            type: object
            properties:
              question_text: { type: string }
              pass_threshold: { type: integer }
        estimated_duration_minutes: { type: integer }
        deadline: { type: string, format: date-time }

    ResearchSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        reward_per_response_idr: { type: number, example: 800 }
        estimated_duration_minutes: { type: integer }
        deadline: { type: string, format: date-time }